[INTERPRETER] * BUG: Final (?) 64 bits fix of the interpreter API. Watching file   descriptors won't crash anymore on 64 bits system when adresses outside   of 32 bits address space are used.
git-svn-id: svn://localhost/gambas/branches/2.0@2304 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gambas.h
@@ -859,7 +859,7 @@ typedef
     void (*MakeDateFromTime)(int, int, GB_DATE *);
     int (*GetTime)(double *, int);
 
-    void (*Watch)(int, int, void *, int);
+    void (*Watch)(int, int, void *, intptr_t);
 
     GB_VALUE *(*Eval)(void *, void *);
